fix the keymaps and what not
