+++ keyboards/cheshire/curiosity/keymaps/default/readme.md
