+++ keyboards/cheshire/curiosity/keymaps/default/rules.mk
