--- keyboards/cheshire/curiosity/keymaps/devinceble/readme.md
@@ -1,9 +1,0 @@
-# Devinceble AKA Vimwarrior ProjectKeyboard Alice
-
-Build Hex File:
-
-    make projectkb/alice:devinceble
-
-Flash Keyboard
-
-    make projectkb/alice:devinceble:flash
