--- keyboards/cheshire/curiosity/keymaps/devinceble/rules.mk
@@ -1,1 +1,0 @@
-MOUSEKEY_ENABLE    = yes
